requirements.txt
anyio==3.5.0
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
dnspython==2.1.0
email-validator==1.1.3
executing==0.8.2
fastapi==0.71.0
google-api-core==2.4.0
google-auth==2.3.3
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
ipython==8.0.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
libcst==0.4.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
motor==2.5.1
mypy-extensions==0.4.3
orjson==3.6.5
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
Pygments==2.11.2
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
stack-data==0.1.4
starlette==0.17.1
tomli==1.2.3
traitlets==5.1.1
typing-inspect==0.7.1
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
